more anki react query refactoring
